Door and Button Added to Scn_Combined
When Player 1 or 2 or the box collide with the door button, the door opens. Problem is still that if two players or a player and a box are colliding with button, and one of them leaves, the door closes. Still needs to be fixed.
Also changed the names of the left and right walls and the ceiling to reflect what they are.